+++ lib/view_component/base.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "action_view"
+require "view_component/cacheable"
 require "active_support/configurable"
 require "view_component/collection"
 require "view_component/compile_cache"
@@ -29,6 +30,7 @@ module ViewComponent
       end
     end
 
+    include ViewComponent::Cacheable
     include ViewComponent::InlineTemplate
     include ViewComponent::UseHelpers
     include ViewComponent::Slotable
@@ -109,13 +111,7 @@ module ViewComponent
 
       if render?
         rendered_template = render_template_for(@__vc_variant, __vc_request&.format&.to_sym).to_s
-
-        # Avoid allocating new string when output_preamble and output_postamble are blank
-        if output_preamble.blank? && output_postamble.blank?
-          rendered_template
-        else
-          safe_output_preamble + rendered_template + safe_output_postamble
-        end
+        __vc_render_cacheable(rendered_template)
       else
         ""
       end
@@ -259,12 +255,6 @@ module ViewComponent
     # @private
     def virtual_path
       self.class.virtual_path
-    end
-
-    # For caching, such as #cache_if
-    # @private
-    def view_cache_dependencies
-      []
     end
 
     # For caching, such as #cache_if