+++ docs/CHANGELOG.md
@@ -22,6 +22,10 @@ nav_order: 5
 
     *Reegan Viljoen*
 
+* Add first class component cache.
+
+    *Reegan Viljoen*
+
 ## 3.21.0
 
 * Updates testing docs to include an example of how to use with RSpec.